☘️ refactor [#12.2.28]: 1차 개선 - 마크다운 렌더러 모듈 분리 및 렌더링 성능 최적화
- **[성능] ReactMarkdown 플러그인 모듈 레벨 상수화**
  - 인라인으로 선언되었던 `remarkPlugins` 및 `rehypePlugins` 배열을 외부 상수로 분리하여, 불필요한 배열 재생성 및 리렌더링을 차단함.

- **[구조/UX] 공통 마크다운 렌더러 팩토리 분리 (`markdown-components.tsx`)**
  - `MessageBubble`에 내장되어 있던 커스텀 마크다운 렌더러를 공용 모듈로 추출.
  - `ChatWindow`의 실시간 스트리밍 렌더링 시에도 동일한 `components={markdownComponents}`를 주입하여, 스트리밍 종료 전/후의 레이아웃 불일치(Layout Shift) 현상을 원천적으로 해결.

🔗 Related:
- Issue [#1047]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/1416#pullrequestreview-4304793120)
